% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_log_related.R
\name{add.to.cleaning.log.other.recode.one}
\alias{add.to.cleaning.log.other.recode.one}
\title{Add existing select_one choices}
\usage{
add.to.cleaning.log.other.recode.one(
  data,
  other_requests,
  is.loop,
  tool.survey,
  tool.choices,
  issue = "Recoding other response",
  label_colname = "label::English",
  existing = "existing.v"
)
}
\arguments{
\item{data}{The dataframe}

\item{other_requests}{The other_requests file, filtered to only contain existing response options}

\item{is.loop}{Whether the data is a loop}

\item{tool.survey}{The survey sheet of the tool}

\item{tool.choices}{The choices sheet of the tool}

\item{issue}{The reason you're recoding this entry, "Recoding other response" by default}

\item{label_colname}{the column with the choice label in your tool.choices. 'label::English' by default}

\item{existing}{your 'existing' column - 'existing.v' by default}
}
\value{
a filled out cleaning log file
}
\description{
This function creates a cleaning log for cases where the 'other' response is already present in the
select_one choices list
}
\note{
This function is not vectorized. It only works 1 row of `other_requests` at a time
}
\examples{
\dontrun{
add.to.cleaning.log.other.recode.one(data = data, other_requests = other_requests, is.loop = F,
tool.survey = tool.survey, tool.choices=tool.choices)
}
}
