% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_log_related.R
\name{add.to.cleaning.log.other.recode}
\alias{add.to.cleaning.log.other.recode}
\title{Add recoded other requests to the cleaning log}
\usage{
add.to.cleaning.log.other.recode(
  data,
  other_requests,
  is.loop,
  tool.survey,
  tool.choices,
  issue = "Recoding other response",
  label_colname = "label::English",
  existing = "existing.v",
  ref.type = "ref.type"
)
}
\arguments{
\item{data}{The dataframe}

\item{other_requests}{Other_requests file}

\item{is.loop}{Whether the dataframe and the request line is a loop}

\item{tool.survey}{The 'survey' sheet of the KOBO tool}

\item{tool.choices}{The 'choices' sheet of the KOBO tool}

\item{issue}{The reason you're recoding this entry, "Recoding other response" by default}

\item{label_colname}{the column with the choice label in your tool.choices. 'label::English' by default}

\item{existing}{your 'existing' column - 'existing.v' by default}

\item{ref.type}{your 'ref.type' column - 'ref.type' by default}
}
\value{
a filled out cleaning log file
}
\description{
Add recoded other requests to the cleaning log
}
\note{
This function is not vectorized. It only works 1 row of `other_requests` at a time.
For the vectorization functionality, please use `vectorized.add.to.cleaning.log.other.recode`
}
\examples{
\dontrun{
add.to.cleaning.log.other.recode(data = data, other_requests = other_requests, is.loop = F,
tool.survey = tool.survey, tool.choices=tool.choices)
}
}
