% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_log_related.R
\name{add.to.cleaning.log.other.remove}
\alias{add.to.cleaning.log.other.remove}
\title{add.to.cleaning.log.other.remove}
\usage{
add.to.cleaning.log.other.remove(
  data,
  other_requests,
  is.loop,
  issue = "Invalid other response"
)
}
\arguments{
\item{data}{your dataframe}

\item{other_requests}{filled out other.requests file (only invalid == T rows) #add a check for this}

\item{is.loop}{whether the data is a loop}

\item{issue}{the reason that the entries were deemed invalid for the cleaning log entry, "Invalid other response" by default}
}
\value{
a filled out cleaning log file for the given entry
}
\description{
The function that takes the filled out other.requests and creates a cleaning log the changes needed for the `invalid` entries
}
\note{
this function is not vectorized - hence it takes on only 1 row of `other_requests` at a time.
If you want to run it over a column, please use `vectorized.add.to.cleaning.log.other.remove`. Just make sure to provide the correct inputs
}
\examples{
\dontrun{
add.to.cleaning.log.other.remove(data = data, other_requests = other.requests, is.loop = F)
}
}
