% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_log_related.R
\name{add.to.cleaning.log.trans.remove}
\alias{add.to.cleaning.log.trans.remove}
\title{Remove invalid responses to open question}
\usage{
add.to.cleaning.log.trans.remove(other_requests, issue = "Invalid response")
}
\arguments{
\item{other_requests}{- a file with invalid answers to open questions}

\item{issue}{- why they were deleted}
}
\value{
a cleaning log for deletion of responses

You can provide it with vectors of invalid answers, it'll work just fine
}
\description{
Remove invalid responses to open question
}
\examples{
\dontrun{
add.to.cleaning.log.trans.remove(other_requests = other.requests)
}
}
