% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_test.R
\name{anomaly_test_enum_diffr}
\alias{anomaly_test_enum_diffr}
\title{Test individual enumerator differences in the audit files}
\usage{
anomaly_test_enum_diffr(
  variable_list,
  audit,
  data_frame,
  enum_name,
  geo_column,
  n_limit = 10,
  sig_thr = 0.05
)
}
\arguments{
\item{variable_list}{- The list of variables to be tested (it is recommended to group questions)}

\item{audit}{- The audit file}

\item{data_frame}{- The main dataframe}

\item{enum_name}{- The name of the column containing enumerator IDs}

\item{geo_column}{- The column in the main dataframe indicating the geographic allocation of the variable}

\item{n_limit}{- The minimum number of interviews the enumerator has to have to be considered}

\item{sig_thr}{- The threshold for the significance filter of the differences test. 0.05 by default}
}
\value{
- a dataframe of problematic enumerators
}
\description{
The function tests for inter-enumerator differences within the distribution of duration answers of enumerators based on
the geographic variable selected.
}
\examples{
\dontrun{
anomaly_test_enum_diffr(variable_list = ls_test,
audit = test_audits,data_frame = test_data,
geo_column = 'oblast',enum_name = 'q0_2_enum_id')
}
}
