% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_test.R
\name{anomaly_test_tmln_int}
\alias{anomaly_test_tmln_int}
\title{Test for timeline anomalies within enumerator audit files}
\usage{
anomaly_test_tmln_int(
  ls,
  audit,
  data,
  geo_column,
  enum_name,
  ipq_n = 0.3,
  n_limit = 10
)
}
\arguments{
\item{ls}{- The list of variables to be tested (it is recommended to group questions)}

\item{audit}{- The audit file}

\item{data}{- The main dataframe}

\item{geo_column}{- The column in the main dataframe indicating the geographic allocation of the variable}

\item{enum_name}{- The name of the column containing enumerator IDs}

\item{ipq_n}{- The percentage of problematic questions that mark an interview as problematic. 0.3 by default}

\item{n_limit}{- The minimum number of interviews the enumerator has to have to be considered}
}
\value{
- a dataframe of problematic enumerators
}
\description{
The function tests for duration outliers within the distribution of duration answers of enumerators based on
the geographic variable selected.
}
\examples{
\dontrun{
anomaly_test_tmln_int(ls = ls_test,audit = test_audits,data = test_data,
geo_column = 'q2_4_1_oblast',enum_name = 'q0_2_enum_id')
}
}
