% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{anychoice_pattern}
\alias{anychoice_pattern}
\title{Collapse a list of multiple choice answers into a more readable format}
\usage{
anychoice_pattern(choices)
}
\arguments{
\item{choices}{- a character variable in your environment}
}
\value{
a single character string composed of the previous vector in a readable format
}
\description{
Collapse a list of multiple choice answers into a more readable format
}
\examples{
choices <- c("a",2,"s","f","a")
anychoice_pattern(choices)
}
