% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_log_related.R
\name{apply.changes}
\alias{apply.changes}
\title{Apply changes to main data basing on a cleaning log.}
\usage{
apply.changes(data, clog, is.loop, print_debug = T)
}
\arguments{
\item{data}{Data (raw.main or raw.loop)}

\item{clog}{Cleaning log - dataframe containing columns uuid, loop_index, variable, new.value, old.value}

\item{is.loop}{Obsolete. This function automatically guesses whether the data contains column 'loop_index'. But there will be warnings produced just in case.}

\item{print_debug}{If True (any by default), warnings related to value mismatches and informational messages about the number of changes will be printed.}
}
\value{
Dataframe containing data with applied changes
}
\description{
Outputs warnings if uuids, loop indeces, or variables from `clog` are not found in `data`. Or if old.value doesn't match what is present in the data
Be aware: all values will be written to data as character.
}
\examples{
\dontrun{
apply.changes(data, clog = c_log, is.loop = F,print_debug = T)
}
}
