% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_stats.R
\name{cf.int}
\alias{cf.int}
\title{Confidence Interval of Mean}
\usage{
cf.int(.x, cf.perc = 0.95, cf.dist = NULL)
}
\arguments{
\item{.x}{Variable to calulate the confidence interval for}

\item{cf.perc}{Confidence interval, i.e. 95\% certainty would be 0.95. Default is 0.95}

\item{cf.dist}{Whether to use a Z-score critical value or a t-score. Default is to check whether there are n > 30 samples in .x, if so a Z-score is used}
}
\description{
Calculate the confidence interval of a variable.
}
