% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{column_cleaner}
\alias{column_cleaner}
\title{Helper function for the find.similar.surveys, can be used separately}
\usage{
column_cleaner(df, labels)
}
\arguments{
\item{df}{Data frame that should be processed}

\item{labels}{list of columns names that should be processed}
}
\description{
Receive data frame and list of columns names that should be processed
make text to lower case -> split by words -> sort by words in ascending order -> combine with " "
}
\examples{
\dontrun{
clean_data <- column_cleaner(data, colnames(data))
}
}
