% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_comparison.R
\name{compare_columns}
\alias{compare_columns}
\title{Compare columns of clean and old dataframes and build a cleaning log}
\usage{
compare_columns(clean_data, raw_data, uuid_col, columns_to_check)
}
\arguments{
\item{clean_data}{Clean data}

\item{raw_data}{Raw dataframe}

\item{uuid_col}{The name of your uuid column}

\item{columns_to_check}{Which columns should be checked}
}
\value{
A cleaning log
}
\description{
Compare columns of clean and old dataframes and build a cleaning log
}
\examples{
\dontrun{
compare_columns(clean_data = clean.data, raw_data = raw.main, uuid_col ='uuid', columns_to_check = col_list)
}
}
