% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_rows}
\alias{compare_rows}
\title{Compare rows of clean and old dataframes and build a deletion log}
\usage{
compare_rows(
  data_raw,
  data_clean,
  id_col,
  col.enum,
  is.loop,
  data.main = NULL,
  reason
)
}
\arguments{
\item{data_raw}{Raw dataframe}

\item{data_clean}{Clean dataframe}

\item{id_col}{The name of your uuid/loop_index column}

\item{col.enum}{The name of the enumerator id column}

\item{is.loop}{Whether the data is a loop}

\item{data.main}{If it's a loop, you'll need to feed it the main dataframe}

\item{reason}{Why they were deleted}
}
\value{
A deletion log
}
\description{
Compare rows of clean and old dataframes and build a deletion log
}
\examples{
\dontrun{
compare_rows(data_raw = raw_data, data_clean = clean_data, id_col = 'uuid', col.enum = 'q0_2_enum_id',is.loop = F,reason = 'test')
}
}
