% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_log_related.R
\name{create.deletion.log}
\alias{create.deletion.log}
\title{Create deletion log}
\usage{
create.deletion.log(data, col_enum, reason, is.loop = F, data.main = NULL)
}
\arguments{
\item{data}{Dataframe containing columns `uuid` and `col_enum`. A subset of `raw.main`, filtered to have the deleted `uuid` `loop_index`}

\item{col_enum}{The name of the column which contains the enumerator's id in your main dataframe}

\item{reason}{This is a string describing the reason for removing a survey from data.}

\item{is.loop}{Whether the `data` argument is a loop or not}

\item{data.main}{If the `data` argument is a loop, you'll need to provide the main dataframe too}
}
\value{
A dataframe containing a deletion log with columns `uuid`, `col_enum`, `reason`, OR an empty dataframe if `data` has 0 rows.
}
\description{
Creates a deletion log for the provided data and reason.
}
\note{
The provision of `data.main` in the case of using `is.loop` = T is needed because we're pulling up the `col_enum`
from the main dataframe.
}
\examples{
\dontrun{
# for non-loops
create.deletion.log(data = data, col_enum = 'enum_col', reason = 'bad entry', is.loop = F)
# for loops
create.deletion.log(data = data_loop, col_enum = 'enum_col', reason = 'bad entry', is.loop = T, data.main = data)
}
}
