% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create.follow.up.requests}
\alias{create.follow.up.requests}
\title{Create a follow-up requests file}
\usage{
create.follow.up.requests(checks.df, directory, wb_name)
}
\arguments{
\item{checks.df}{The checks file for the follow-up requests}

\item{directory}{The directory for your follow-up request}

\item{wb_name}{The name of your workbook}
}
\value{
Write a follow-up requests file
}
\description{
Create a follow-up requests file
}
\examples{
\dontrun{
create.follow.up.requests(checks.df = data, wb_name = 'other_requests', directory = 'output/checking/requests/')
}
}
