% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create.translate.requests}
\alias{create.translate.requests}
\title{Format a translate responses file}
\usage{
create.translate.requests(responses.j, response_colname = "response.en")
}
\arguments{
\item{responses.j}{Dataframe containing responses to any questions from `questions.db`}

\item{response_colname}{String containing name of the column which has the relevant response (tanslated to English in most cases) unused}
}
\value{
Returns a formatted recode.others dataframe
}
\description{
Format a dataframe containing responses to prepare for other/translate requests
}
\details{
Relocates columns and adds the TEI columns.
}
\examples{
\dontrun{
create.translate.requests(responses.j = responses.j, response_colname = "response.en")
}
}
