% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{find.responses}
\alias{find.responses}
\title{Look up a raw dataframe to find all responses to a given set of questions.}
\usage{
find.responses(.dataframe, questions.db, values_to = "responses", is.loop = F)
}
\arguments{
\item{.dataframe}{raw data}

\item{questions.db}{dataframe that contain a column `name` such as a subset of `tool.survey`}

\item{values_to}{Name of th vector containing the found responses}

\item{is.loop}{A boolean defining if the raw data you are using is the main or one of the loops}
}
\value{
A dataframe containing the uuid, name and the column specified by values_to.
        In case is.loop is T, then a column called loop_index will be added with the unique values.
}
\description{
Look up a raw dataframe to find all responses to a given set of questions.
}
\examples{
q.db <- data.frame(name = c("age","occupation"), ref.name = c("age_choice", "occupation_choice"))
raw.data <- data.frame(age = c(21,32), occupation = c("cook","train conductor"), uuid = c("abc","def"),
                      age_choice = c("other", "other"), occupation_choice = c("other", "other"))
res <-find.responses(raw.data, q.db, "responses")
}
