% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find.similar.surveys}
\alias{find.similar.surveys}
\title{For each survey, it finds the closest matching survey with the same enumerator
and with the minimum number of different column}
\usage{
find.similar.surveys(
  data.main,
  tool.survey,
  uuid = "_uuid",
  idnk_value = "idnk",
  enum.column = NA,
  is.loop = FALSE
)
}
\arguments{
\item{data.main}{DataFrame with raw data}

\item{tool.survey}{tool.survey data frame}

\item{uuid}{Name of the column in which uuids are stored.}

\item{idnk_value}{Value (from tool.choices) that represents the answer "I don't know"}

\item{enum.column}{Name of column that represent unique identifier for the enumerators}

\item{is.loop}{If TRUE, then calculate questions preserved}
}
\description{
For each survey, it finds the closest matching survey with the same enumerator
and with the minimum number of different column
}
\details{
- Uses gower distance when finding the most similar surveys
- Searches for similar survays in the section of the same enumerator, namely the column specified in enum.column
}
\examples{
\dontrun{
outdata <- find.similar.surveys(
 data.main=raw.main,
 tool.survey=tool.survey,
 enum.column='a2_1_enum_id',
 is.loop=F)
}
}
