% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.choice.label}
\alias{get.choice.label}
\title{Find the label of the Choice}
\usage{
get.choice.label(choice, list, label_colname = NULL, tool.choices = NULL)
}
\arguments{
\item{choice}{the name of the choice}

\item{list}{the name of the list containing choice}

\item{label_colname}{This is the label_colname input}

\item{tool.choices}{This is the tool.choices data.frame}
}
\value{
It will return the label of the choice
}
\description{
Find the label of the Choice
}
\examples{
\dontrun{
choice_label <- get.choice.label(choice = "yes", list = "yn",
                                 label_colname = label_colname,
                                 tool.choices = tool.choices)
}
}
