% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.choice.list.from.type}
\alias{get.choice.list.from.type}
\title{Finds the choice list for a question basing on its type}
\usage{
get.choice.list.from.type(q_type)
}
\arguments{
\item{q_type}{The type of the variable from the kobo tool}
}
\value{
the list_name of the choices
}
\description{
Finds the choice list for a question basing on its type
}
\examples{
\dontrun{
  choice_list_name <- get.choice.list.from.type("select_one a2_partner")
}
}
