% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.choice.name.from.label}
\alias{get.choice.name.from.label}
\title{Find the name of the Choice from label}
\usage{
get.choice.name.from.label(
  choice_label,
  list,
  label_colname = NULL,
  tool.choices = NULL
)
}
\arguments{
\item{choice_label}{the label of the choice}

\item{list}{the name of the list containing choice}

\item{label_colname}{This is the label_colname input}

\item{tool.choices}{This is the tool.choices data.frame}
}
\value{
It will return the name of the choice label
}
\description{
Find the name of the Choice from label
}
\examples{
\dontrun{
choice_name <- get.choice.name.from.label(choice = "Yes", list = "yn",
                                 label_colname = label_colname,
                                 tool.choices = tool.choices)
}
}
