% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.label}
\alias{get.label}
\title{Find the label of a variable}
\usage{
get.label(variable, label_colname = NULL, tool.survey = NULL)
}
\arguments{
\item{variable}{This is the name of the header from raw data.}

\item{label_colname}{This is the label_colname input}

\item{tool.survey}{This is the tool.survey data.frame}
}
\value{
It will return the label of the question from the kobo tool
}
\description{
Find the label of a variable
}
\examples{
\dontrun{
label <- get.label(variable = "a2_partner", label_colname = label_colname, tool.survey = tool.survey)
}
}
