% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.other.db}
\alias{get.other.db}
\title{Finds all 'other' questions and their ref question and choices}
\usage{
get.other.db(tool.survey = NULL, tool.choices = NULL, label_colname = NULL)
}
\arguments{
\item{tool.survey}{This is the tool.survey data.frame}

\item{tool.choices}{This is the tool.choices data.frame}

\item{label_colname}{This is the label_colname input}
}
\value{
Dataframe containing name, parent name, full.label, parent type, choices, choices label
}
\description{
Finds all 'other' questions and their ref question and choices
}
\examples{
\dontrun{
other_db <- get.other.db(tool.choices = tool.choices,
                         tool.survey = tool.survey,
                         label_colname = label_colname)
}

}
