% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.ref.question}
\alias{get.ref.question}
\title{Finds parent question basing on relevancy text}
\usage{
get.ref.question(q_relevancy)
}
\arguments{
\item{q_relevancy}{The relevancy entry from the kobo tool}
}
\value{
the parent questions variable name
}
\description{
Finds parent question basing on relevancy text
}
\examples{
\dontrun{
parent_variable <- get.ref.question("selected(${b17_access_stores}, 'other'"))
}

}
