% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.select.db}
\alias{get.select.db}
\title{Finds all 'select' type questions, and their choices}
\usage{
get.select.db(tool.choices = NULL, tool.survey = NULL, label_colname = NULL)
}
\arguments{
\item{tool.choices}{This is the tool.choices data.frame}

\item{tool.survey}{This is the tool.survey data.frame}

\item{label_colname}{This is the label_colname input (usually `label::English``)}
}
\value{
a data.frame including all select questions with the type, name, label, q.type, list_name, and choices
}
\description{
Finds all 'select' type questions, and their choices
}
\examples{
\dontrun{
select_db <- get.select.db(tool.choices = tool.choices,
                           tool.survey = tool.survey,
                           label_colname = label_colname)
}

}
