% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.text.db}
\alias{get.text.db}
\title{Finds all text questions}
\usage{
get.text.db(tool.survey = NULL, tool.choices = NULL, label_colname = NULL)
}
\arguments{
\item{tool.survey}{This is the tool.survey data.frame}

\item{tool.choices}{This is the tool.choices data.frame}

\item{label_colname}{This is the label_colname input (usually `label::English``)}
}
\value{
A dataframe of all text questions in your survey
}
\description{
Finds all text questions
}
\note{
This function exists for cases where the person who coded the Kobo questionnaire didn't follow the conventional rules
For writing them and some of the questions were ommited from other `other.db` or `trans.db`. This function will create an
object that is compatible with both of those dataframes and relevant rows from it can be appened to either `other.db`
or `trans.db`.
}
\examples{
\dontrun{
text.db <- get.text.db(tool.choices = tool.choices,
                         tool.survey = tool.survey,
                         label_colname = label_colname)
}
}
