% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.trans.db}
\alias{get.trans.db}
\title{Finds all questions which should be translated (meaning all 'text'-type question that are not 'other's)}
\usage{
get.trans.db(tool.survey = NULL, tool.choices = NULL, label_colname = NULL)
}
\arguments{
\item{tool.survey}{This is the tool.survey data.frame}

\item{tool.choices}{This is the tool.choices data.frame}

\item{label_colname}{This is the label_colname input}
}
\value{
Dataframe containing name,full.label
}
\description{
Finds all questions which should be translated (meaning all 'text'-type question that are not 'other's)
}
\examples{
\dontrun{
trans_db <- get.trans.db(tool.choices = tool.choices,
                         tool.survey = tool.survey,
                         label_colname = label_colname)
}
}
