% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.type}
\alias{get.type}
\title{Find the type of variables.}
\usage{
get.type(variable, tool.survey = NULL)
}
\arguments{
\item{variable}{This is the name of the header from raw data.}

\item{tool.survey}{This is the tool.survey data.frame}
}
\value{
It will return the question type from the kobo tool
}
\description{
Find the type of variables.
}
\examples{
\dontrun{
question_type <- get.type(variable = "a2_partner",tool.survey = tool.survey)
}
}
