% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%_>_\%}
\alias{\%_>_\%}
\title{Check if a is greater than b}
\usage{
a \%_>_\% b
}
\arguments{
\item{a}{- a variable present in the environment}

\item{b}{- a variable present in the environment}
}
\value{
- A boolean for the check
}
\description{
Check if a is greater than b
}
\examples{
a <- 3
a \%_>_\% 2
}
