% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%_<=_\%}
\alias{\%_<=_\%}
\title{Check if a is less than or equal to b}
\usage{
a \%_<=_\% b
}
\arguments{
\item{a}{- a numeric variable present in the environment}

\item{b}{- a numeric variable present in the environment}
}
\value{
- A boolean for the check
}
\description{
Check if a is less than or equal to b
}
\examples{
a <- 5
a \%_<=_\% 7
}
