% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%_+_\%}
\alias{\%_+_\%}
\title{Produce a sum of two numeric variables}
\usage{
a \%_+_\% b
}
\arguments{
\item{a}{- a variable present in the environment}

\item{b}{- a variable present in the environment}
}
\value{
- A numeric sum
}
\description{
Produce a sum of two numeric variables
}
\examples{
a <- 4
a \%_+_\% 2
}
