% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%_*_\%}
\alias{\%_*_\%}
\title{Produce a result of multiplication of two numeric variables}
\usage{
a \%_*_\% b

a \%_*_\% b
}
\arguments{
\item{a}{- a variable present in the environment}

\item{b}{- a variable present in the environment}
}
\value{
- a numeric result of multiplication

- a numeric result of multiplication
}
\description{
Produce a result of multiplication of two numeric variables

Produce a result of multiplication of two numeric variables
}
\examples{
a <- 4
a \%_*_\% 2

a <- 4
a \%_*_\% 2

}
