% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%==\%}
\alias{\%==\%}
\title{Equality checker}
\usage{
a \%==\% b
}
\arguments{
\item{a}{- a numeric/string variable present in the environment}

\item{b}{- a numeric/string variable present in the environment}
}
\value{
- A boolean for the equality check
}
\description{
Equality checker
}
\examples{
a <- 2
a\%==\%2

}
