% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%!=\%}
\alias{\%!=\%}
\title{Inequality checker}
\usage{
a \%!=\% b
}
\arguments{
\item{a}{- a numeric/string variable present in the environment}

\item{b}{- a numeric/string variable present in the environment}
}
\value{
- A boolean for the inequality check
}
\description{
Inequality checker
}
\examples{
a <- 3
a \%!=\% 2
}
