% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%not=na\%}
\alias{\%not=na\%}
\title{Check if two variables are non-equal and not NA}
\usage{
e1 \%not=na\% e2
}
\arguments{
\item{e1}{- a variable present in the environment}

\item{e2}{- a variable present in the environment}
}
\value{
- a boolean result of the comparison
}
\description{
Check if two variables are non-equal and not NA
}
\examples{
a <- 4
b <- 5
a \%not=na\% b
}
