% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_data_man.R
\name{grext}
\alias{grext}
\title{Base Stringr}
\usage{
grext(text, pattern)
}
\arguments{
\item{text}{Text string to have matched}

\item{pattern}{Text string to match in text}
}
\description{
Stringr functions rewritten to use only base functions
}
\details{
\code{grext} is the equivalent of \code{stringr::str_extract}
}
