% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_data_man.R
\name{grextall}
\alias{grextall}
\title{Base Stringr}
\usage{
grextall(text, pattern, simplify = FALSE)
}
\arguments{
\item{text}{Text string to have matched}

\item{pattern}{Text string to match in text}

\item{simplify}{Whether to return an unlisted character vector, default
is \code{FALSE} and returns list}
}
\description{
Stringr functions rewritten to use only base functions
}
\details{
\code{grextall} is the equivalent of \code{stringr::str_extract_all}
}
