% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_data_man.R
\name{guess_names}
\alias{guess_names}
\title{Data Ingest}
\usage{
guess_names(.data, sensitivity = 1, clean_names = TRUE)
}
\arguments{
\item{.data}{A \code{data.frame} that needs clean names}

\item{sensitivity}{A numeric indicating how many name repaired column names to allow before picking column headers}

\item{clean_names}{A boolean, whether to return the \code{data.frame} with cleaned names from \code{janitor} package}
}
\description{
Find the correct row to use as column names without needing
to specify \code{read_excel(..., skip = 3)}. Simply read in data and pipe
here.
}
