% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{isna}
\alias{isna}
\title{A shorthand is NA function}
\usage{
isna(x)
}
\arguments{
\item{x}{- a variable present in the environment}
}
\value{
- a boolean of whether X was NA
}
\description{
A shorthand is NA function
}
\examples{
a <- 2
isna(2)
}
