% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_data_man.R
\name{joinCoalesce}
\alias{joinCoalesce}
\title{Joining Data}
\usage{
joinCoalesce(x, y, join, by = NULL, suffix = c(".x", ".y"), ignore = NULL)
}
\arguments{
\item{x}{a \code{data.frame} object}

\item{y}{a \code{data.frame} object}

\item{join}{a \code{dplyr} join function to apply to \code{x} and \code{y}.}

\item{suffix}{suffixes to apply to column names found in both \code{x} and \code{y}. Columns with these suffixes will be coalesced together.}

\item{ignore}{character vector of column names that exist in both \code{x} and \code{y} that shouldn't be coalesced.}
}
\description{
Join together two data sets and coalesce together columns with
the same name that aren't used as joining columns.
}
\examples{
\dontrun{
one <- tribble(
      ~x,~y,~z,
      "a", 1,1,
      "b", 9,1,
      "c", 4,1
      )

two <- tribble(
     ~x, ~y,~z,
     "a", NA,2,
     "b",2,2,
     "d", 9,2,
     "c", NA,2
      )
joinCoalesce(one, two, dplyr::left_join, by = "x", ignore = "z")
}
}
