% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.audit.files}
\alias{load.audit.files}
\title{Load Audit files}
\usage{
load.audit.files(dir.audits, uuids = NULL, track.changes = F, add.uuid2 = T)
}
\arguments{
\item{dir.audits}{The directory in which to look for audit files (path resembling .../data/audits/...)}

\item{uuids}{The uuids of surveys that are to be loaded. If NULL is provided here (and by default) all uuids from dir.audits will be loaded.}

\item{track.changes}{Whether the survey has the parameter track-changes set to `true`}

\item{add.uuid2}{Whether the function should add uuid2. If TRUE, this'll allow the user a better control over error messages in proccess uuid}
}
\value{
Returns a dataframe with contents of all `audit.csv` files from `dir.audits` or its subdirectories.
}
\description{
Loads all audit files from a selected folder into 1 large dataframe
}
\examples{
\dontrun{
load.audit.files(dir.audits = 'your_path', uuids = data$uuid)
}
}
