% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.tool_colname.R
\name{load.label_colname}
\alias{load.label_colname}
\title{Load the label column name from the tool survey}
\usage{
load.label_colname(filename_tool, language = "English")
}
\arguments{
\item{filename_tool}{a link to the tool filename}

\item{language}{define the language of the label to be used}
}
\value{
the name of the label column
}
\description{
It is better to assign this function to a variable called label_colname.
This variable will be the input to many other functions.
}
\examples{
\dontrun{
filename_tool <- "tool.xlsx"
label_colname <- load.label_colname(filename_tool, language = "English")
}
}
