% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.requests}
\alias{load.requests}
\title{Load 'request' logs from specified directory.}
\usage{
load.requests(dir, filename.pattern, sheet = NULL, validate = FALSE)
}
\arguments{
\item{dir}{Directory which should be searched for files.}

\item{filename.pattern}{String with a regex pattern which will be passed to `list.files` to match files,
however: '^' (string start) is added at the start of the pattern, and ".*\\.xlsx" is added at the end,
so effectively files that will be loaded must be XLSX and have names that start with the provided pattern.}

\item{sheet}{Optional parameter passed to `read_xlsx`, defaults to NULL (first sheet of an Excel workbook)}

\item{validate}{Should the file be validated (make sure that only one of TEI columns is filled.)}
}
\value{
Your request file in the tibble format
}
\description{
Searches `dir` to find XLSX files with names that start with a match for `filename.pattern`.
NB: This pattern-matching is case-insensitive. If files contain the classic "TRUE", "EXISTING" or "INVALID" (TEI) columns,
these will be renamed to "true.v", "existing.v", and "invalid.v" respectively and optionally validated for errors.
}
\examples{
\dontrun{
load.requests(dir = 'your_path', filename.pattern = 'your_pattern')
}
}
