% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.tool.survey}
\alias{load.tool.survey}
\title{Load the 'survey' tab from a Kobo tool.}
\usage{
load.tool.survey(filename_tool, label_colname, keep_cols = F)
}
\arguments{
\item{filename_tool}{This is the path to the file that contains the tool}

\item{label_colname}{This is the variable of the label colname}

\item{keep_cols}{Whether all columns from the original tool should be kept. By default it is False, meaning that only the relevant label, hints, etc are kept}
}
\value{
A dataframe of the survey from the kobo tool, with new columns added: `datasheet`, `q.type`, and `list_name`
}
\description{
Load the 'survey' tab from a Kobo tool.
}
\examples{
\dontrun{
filename_tool <- "tool.xlsx"
label_colname <- load.label.colname(filename_tool)
tool_survey <- load.tool.survey(filename_tool, label_colname, keep_cols = F)
}
}
