% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_comparison.R
\name{lookup_columns}
\alias{lookup_columns}
\title{Look up the values of specified columns in 2 different datasets}
\usage{
lookup_columns(id, id_column, column, clean.data, raw.data)
}
\arguments{
\item{id}{The ids you want to look up}

\item{id_column}{The name of the ID column}

\item{column}{The column you want to look up}

\item{clean.data}{Your clean dataset}

\item{raw.data}{Your raw dataset}
}
\value{
A lookup table comparing the old and new values of a column for the given uuid
}
\description{
Look up the values of specified columns in 2 different datasets
}
\examples{
\dontrun{
lookup_columns(id = 'uuid_example', id_column = 'uuid', column = 'q_1_a',
clean.data = clean.main, raw.data = raw.main)
}
}
