% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_log_related.R
\name{make.logical.check.entry}
\alias{make.logical.check.entry}
\title{Create a logical check DF}
\usage{
make.logical.check.entry(
  check,
  id,
  question.names,
  issue,
  cols_to_keep = c(),
  is.loop,
  date = NA
)
}
\arguments{
\item{check}{Dataframe with data, filtered according to some flag (variable + uuid/loop_index filter).
Must contain columns `uuid` and all columns in `question.names`}

\item{id}{The identifier of this logical check.}

\item{question.names}{List of relevant queston names for this logical check.}

\item{issue}{description of the issue encountered}

\item{cols_to_keep}{List of columns from raw.main to be included in result.}

\item{is.loop}{Obsolete. This function automatically guesses whether the data contains column 'loop_index'.
But there will be warnings produced just in case.}

\item{date}{If you want to include the `date` column, you can specify it's name in here}
}
\value{
Dataframe containing at the least columns: `uuid`, `check.id`, `variable`, `issue`, `old.value`, `new.value`, `explanation`.
This object can be later added to cleaning log.
}
\description{
Create a logical check DF
}
\examples{
\dontrun{
make.logical.check.entry(check = data, id = 1, question_names = c('b11_gas_heating_price','b10_gas_vehicle_price'), issue = 'test', is.loop = F)
}
}
