% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_data_man.R
\name{map_fns}
\alias{map_fns}
\title{Pipe Alt. - Mapping Functions}
\usage{
map_fns(.x, .fns, error_handling = "do-nothing")
}
\arguments{
\item{.x}{An object to apply functions to}

\item{.fns}{A \code{list} of functions to apply in sequence to \code{.x}. E.g. list(toupper, function(x) {str_remove(x, " ")})}

\item{error_handling}{How errors in code should be handled. Options are \code{"do-nothing"} and \code{"stop"}. \code{"do-nothing"} skips the function in the sequence that produced the error and continues through the list. \code{"stop"} results in the code stopping with standard error messaging.}
}
\description{
Applies a list of functions to an object in sequential order. Usefule in cases where you don't want to use a lot of \code{\%>\%} operators. Usefule in conjunction with \code{\%=\%} and \code{g}.
}
\examples{
\dontrun{
map_fns("THe Starting price is $25,000 to $35,000", list(function(x) {str_extract_all(x, "\\\\$[0-9,]+")}, unlist, function(x){str_remove_all(x, "[^0-9]")}, as.numeric, function(x){mean(x, na.rm = TRUE)}))

# Using with \%=\% and g

# cleaning steps

cln_lst <- list(toupper, str_squish, function(x) {str_remove_all(x, "[^A-Z]")})
make <- "Chevro0lT"
model <- "odYssey5"
g(make, model) \%=\% purrr::map(c(make, model), ~ map_fns(.x, cln_lst))
}
}
