% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_data_man.R
\name{na_if_func}
\alias{na_if_func}
\title{Na if extension}
\usage{
na_if_func(x, .f, ...)
}
\arguments{
\item{x}{object that contains values to be switched to NA}

\item{.f}{a function that returns a boolean, such as \code{stringr::str_detect}}
}
\description{
Extends \code{na_if}, convert data to \code{NA} if behavior is
detected. Useful in mutates and long pipes. Equivalent to \code{x[f(x)]}.
}
\examples{
\dontrun{
mutate(mtcars, mpg2 = na_if_func(mpg, function(x) x > 30))
}
}
