% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_extension_cords.R
\name{parallel_longer}
\alias{parallel_longer}
\title{Pivot Longer - Easy Parallel Pivot}
\usage{
parallel_longer(
  .data,
  cols,
  parallel = NULL,
  val_names = NULL,
  id_names = NULL
)
}
\arguments{
\item{.data}{A \code{data.frame} with columns to pivot into longer format}

\item{cols}{A character vector of columns to pivot into longer format, these will be used to
group the parallel columns}

\item{parallel}{A list of column groups to pivot into longer format in parallel with \code{cols}.
Each group should be the same length as \code{cols}. When pivoted, the
first column in \code{cols} matches to the first column in each group in \code{parallel}. Default is
\code{NULL} and results in behavior equivalent to \code{tidyr::pivot_longer}.}

\item{val_names}{Names to give the value columns of the resulting dataframe}

\item{id_names}{Names to give the id columns of the resulting dataframe}
}
\description{
Easiliy pivot multiple \code{data.frame} columns into longer format in
parallel
}
\examples{
\dontrun{
parallel_longer(mtcars, c("mpg", "cyl"), list(c("cyl", "hp")), c("VALONE", "VALTWO"), c("COL1", "COL2"))
}
}
