% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_data_man.R
\name{package.dependencies}
\alias{package.dependencies}
\alias{install.dependencies}
\title{Package-Dependencies}
\usage{
package.dependencies(package, level = c("Depends", "Imports", "Suggests"))

install.dependencies(
  package,
  level = c("Depends", "Imports", "Suggests"),
  record = FALSE
)
}
\arguments{
\item{package}{Name of the package to look for dependencies for, as a character string}

\item{level}{Character vector of the levels of dependencies to install, options are
\code{c("Depends", "Imports", "Suggests")}.}

\item{record}{Whether to recrd the newly installed package in your \code{renv.lock} file}
}
\description{
Ever run into cases where package installation fails repeatedly,
with a string of error messages for one package requiring different versions
of different packages? Get a list of all dependencies with \code{package.dependencies}
and install them all in one go with \code{install.dependencies}
}
