% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{pre.process.audits}
\alias{pre.process.audits}
\title{Pre-processing audit files}
\usage{
pre.process.audits(df, threshold)
}
\arguments{
\item{df}{Audit files loaded with `load.audit.files` function}

\item{threshold}{Your numeric threshold for}
}
\value{
Audit files without the outlier duration values
}
\description{
In cases when you believe that the enumerator was stalling on some of the questions to increase the time of the
interview, you can fix it with this function. Every value of the `duration` column of your audits above your
`threshold` value will be set to the question group average calculated by omitting the values above this threshold.
If there are no such values, the value for such rows will be set to 0.
}
\examples{
\dontrun{
pre.process.audits(df, 10)
}
}
