% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process.audit.geospatial}
\alias{process.audit.geospatial}
\title{Geospatial processing of audit data}
\usage{
process.audit.geospatial(df, start_q, end_q)
}
\arguments{
\item{df}{the dataframe of audits}

\item{start_q}{the question you consider to be the start of the interview}

\item{end_q}{the question you consider to be the end of the interview}
}
\value{
a dataframe containing the id, time and location data for the first, second and last set of coodrinates per interview
}
\description{
Geospatial processing of audit data
}
\note{
This function is non vectorized. Mostly used in pipes after a `group_by(uuid)` command
}
\examples{
\dontrun{
loaded.audits \%>\%
  dplyr::group_by(uuid) \%>\%
  dplyr::group_modify(~process.audit.geospatial(.x)) \%>\%
  dplyr::ungroup()
}
}
