% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_data_man.R
\name{Multiple-Assignment}
\alias{Multiple-Assignment}
\alias{\%=\%}
\alias{\%=\%.lbunch}
\title{assign multiple results to multiple variables}
\usage{
l \%=\% r

`\%=\%.lbunch`(l, r, ...)
}
\description{
assign multiple results to multiple variables
}
