% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.multiple.add.choices}
\alias{recode.multiple.add.choices}
\title{Create a cleaning log that add particular choices to the select_multiple variables.}
\usage{
recode.multiple.add.choices(data, variable, choices, issue)
}
\arguments{
\item{data}{Dataframe containing records which will be affected.}

\item{variable}{String containing the name of the select_multiple variable.}

\item{choices}{Vector of strings (or a single string) containing the choices that will be added.
They must be valid options for this variable.}

\item{issue}{String with explanation used for the cleaning log entry.}
}
\value{
Dataframe containing cleaning log entries constructed from `data`.
}
\description{
Changes all 0s to 1s in choice columns specified by `choices`. Modifies cumulative variable too.
}
\note{
This function does not affect entries that have NA in `variable`.
}
\examples{
\dontrun{
recode.multiple.add.choices(data = filter(raw.main, condition), variable = "question_name",
choices = c("option1", "option2"), issue = "explanation")
}
}
