% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.multiple.set.choices}
\alias{recode.multiple.set.choices}
\title{Create a cleaning log that sets the select_multiple variables to one particular choice.}
\usage{
recode.multiple.set.choices(
  data,
  variable,
  choices,
  issue,
  other_var_name = NULL,
  tool.survey,
  tool.choices
)
}
\arguments{
\item{data}{Dataframe containing records which will be affected.}

\item{variable}{String containing the name of the select_multiple variable.}

\item{choices}{Vector of strings (or a single string) containing the choices that will be added. They must be valid options for this variable.}

\item{issue}{String with explanation used for the cleaning log entry.}

\item{other_var_name}{the name of your _other variable - variable containing string responses, NULL by default}

\item{tool.survey}{your survey sheet in the tool}

\item{tool.choices}{your choices sheet in the tool}
}
\value{
Dataframe containing cleaning log entries constructed from `data`.
}
\description{
This function affects choice columns: setting 1 for `choices` and 0 everywhere else.
For the cumulative variable, the value will be more or less equal to `paste(choices, collapse=" ")`
(to be precise, the exact order of choices will be the same as the one in `tool.choices`) ...
Additionally, the _other variable will be set to NA if it is not one of the `choices`
}
\examples{
\dontrun{
recode.multiple.set.choices(data = filter(raw.main, condition),
variable = "question_name", choices = "option", issue = "explanation")
}
}
