% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.multiple.set.NA}
\alias{recode.multiple.set.NA}
\title{Create a cleaning log that sets the select_multiple responses to NA.}
\usage{
recode.multiple.set.NA(
  data,
  variable,
  issue,
  other_var_name = NULL,
  CL_COLS = c("uuid", "uniqui", "loop_index", "variable", "old.value", "new.value",
    "issue")
)
}
\arguments{
\item{data}{Dataframe containing records which will be affected.}

\item{variable}{String containing the name of the select_multiple variable.}

\item{issue}{String with explanation used for the cleaning log entry.}

\item{other_var_name}{the name of the variable containing the "other" responses}

\item{CL_COLS}{string of columns of the cleaning log.
Defauls to c("uuid","uniqui", "loop_index", "variable", "old.value", "new.value", "issue")}
}
\value{
Dataframe containing cleaning log entries constructed from `data`.
}
\description{
Changes all 1s and 0s to NA in choice columns, sets cumulative variable and _other text answers to NA.
}
\examples{
\dontrun{
recode.multiple.set.NA(data = filter(raw.main, condition),
 variable = "question2")
  }
}
