% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.other.relevances}
\alias{recode.other.relevances}
\title{Create a cleaning log for relevant questions from your cleaning log}
\usage{
recode.other.relevances(
  cleaning.log.other,
  relevancy_dictionary,
  data,
  is.loop
)
}
\arguments{
\item{cleaning.log.other}{- the cleaning log that you've got after running recode.others}

\item{relevancy_dictionary}{- the dictionary of relevancies from find.relevancies function}

\item{data}{- your dataframe}

\item{is.loop}{- whether your data is a loop}
}
\value{
a cleaning log for the relevancies of your questions
}
\description{
Some of the variables we're recoding have relevancies that only open up if a choice is answered. Now,
since we're recoding some of those choices to be already existing or invalid, we need to change the answers
for the relevancy as well. This function does just that
}
\examples{
\dontrun{
recode.other.relevances(cleaning.log.other = cleaning.log, relevancy_dictionary = dictionary,
data = raw.main, is.loop = F)
}
}
