% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.others.elsewhere}
\alias{recode.others.elsewhere}
\title{Recode elsewhere others for select multiple and select one questions}
\usage{
\method{recode}{others.elsewhere}(data, tool.survey, or.edited, is.loop)
}
\arguments{
\item{data}{Dataframe containing Kobo data}

\item{tool.survey}{The survey sheet of your kobo tool}

\item{or.edited}{The file with the translated requests loaded with the `load.requests` function}

\item{is.loop}{Logical, whether the data is looped or not}
}
\value{
Returns a cleaning log file
}
\description{
Recode elsewhere others for select multiple and select one questions
}
\examples{
\dontrun{
recode.others.elsewhere(data = raw.main, tool.survey = tool.surveyor.edited = or.edited, is.loop = F)
}
}
