% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.others_select_multiple}
\alias{recode.others_select_multiple}
\title{Create a cleaning log for recoding select_multiple questions}
\usage{
recode.others_select_multiple(
  data,
  or.select_multiple,
  orig_response_col = "response.uk",
  print_debug = T,
  is.loop,
  tool.survey_others = tool.survey,
  tool.choices_others = tool.choices,
  label_colname = "label::English",
  CL_COLS = c("uuid", "uniqui", "loop_index", "variable", "old.value", "new.value",
    "issue"),
  none_selection = c("do_not_know", "prefer_not_to_answer", "none", "none_of_the_above",
    "dont_know", "do_not_want_to_answer")
)
}
\arguments{
\item{data}{your dataframe}

\item{or.select_multiple}{Your filled out recoding form after translation and verification (with only select multiple)}

\item{orig_response_col}{The name of your original column with the untranslated responses, "response.uk" by default}

\item{print_debug}{Whether you want the function to print out what it's doing}

\item{is.loop}{Whether you're working with loop data}

\item{tool.survey_others}{The survey sheet of your kobo tool}

\item{tool.choices_others}{The choices sheet of your kobo tool}

\item{label_colname}{The name of your english label column. "label::English" by default}

\item{CL_COLS}{CL_COLS string of columns of the cleaning log.
Defauls to c("uuid","uniqui" ,"loop_index", "variable", "old.value", "new.value", "issue")}

\item{none_selection}{The `name` values of the columns that make all other answers in a select multiple invalid.
If you're recoding a value in the `existing.v` column into one of these, all other replies for this question will be
removed. If the correct response is `none`, other responses can't also be present. Change depending on your research cycle,
c('do_not_know','prefer_not_to_answer','none','none_of_the_above','dont_know','do_not_want_to_answer') by default}
}
\value{
A dataframe for the changes that need to be applied to the data
}
\description{
Create a cleaning log for recoding select_multiple questions
}
\note{
When running loop data, make sure to have a uuid column in your dataframe.
This ensures that the resulting tables are uniform
}
\examples{
\dontrun{
recode.others_select_multiple(data, or.select_one = other_requests_file, tool.survey_others=tool.survey,
tool.choices_others = tool.choices)
}
}
