% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.others_select_one}
\alias{recode.others_select_one}
\title{Create a cleaning log for recoding select_one questions}
\usage{
recode.others_select_one(
  or.select_one,
  orig_response_col = "response.uk",
  print_debug = T,
  tool.survey_others = tool.survey,
  tool.choices_others = tool.choices,
  label_colname = "label::English",
  CL_COLS = c("uuid", "uniqui", "loop_index", "variable", "old.value", "new.value",
    "issue")
)
}
\arguments{
\item{or.select_one}{Your filled out recoding form after translation and verification  (with only select ones)}

\item{orig_response_col}{The name of your original column with the untranslated responses, "response.uk" by default}

\item{print_debug}{Whether you want the function to print out what it's doing}

\item{tool.survey_others}{The survey sheet of your kobo tool}

\item{tool.choices_others}{The choices sheet of your kobo tool}

\item{label_colname}{The name of your english label column. "label::English" by default}

\item{CL_COLS}{string of columns of the cleaning log. Defauls to c("uuid","uniqui", "loop_index", "variable", "old.value", "new.value", "issue")}
}
\value{
A dataframe for the changes that need to be applied to the data
}
\description{
Create a cleaning log for recoding select_one questions
}
\note{
_others suffix is added to the kobo tool dataframes to avoid recursion
}
\examples{
\dontrun{
recode.others_select_one(or.select_one = other_requests_file, tool.survey_others=tool.survey,
tool.choices_others = tool.choices)
}
}
