% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.set.NA.if}
\alias{recode.set.NA.if}
\title{Recode a question by setting variables to NA if they are equal to a given value (code).}
\usage{
recode.set.NA.if(
  data,
  variables,
  code,
  issue,
  ignore_case = T,
  CL_COLS = c("uuid", "loop_index", "variable", "old.value", "new.value", "issue")
)
}
\arguments{
\item{data}{Dataframe containing records which will be affected.}

\item{variables}{Vector of strings (or a single string) containing the names of the variables.}

\item{code}{Vector of strings (or a single string) which will be changed to NA.}

\item{issue}{String with explanation used for the cleaning log entry.}

\item{ignore_case}{Whether `code` should be matched case-insensitively. Defaults to True.}

\item{CL_COLS}{string of columns of the cleaning log defauls to c("uuid", "loop_index", "variable", "old.value", "new.value", "issue")}
}
\value{
Dataframe containing cleaning log entries constructed from `data`.
}
\description{
Recode a question by setting variables to NA if they are equal to a given value (code).
}
\note{
DO NOT use this function for select_multiple questions. Instead use `recode.multiple.set.NA`
}
\examples{
\dontrun{
 recode.set.NA.if(data = filter(raw.main, condition),
 variables = c("question1", "question2"),
 code = "999", issue = "explanation")
 }
}
