% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.set.NA}
\alias{recode.set.NA}
\title{Set the given variables for the given entries to NA.}
\usage{
recode.set.NA(data, variables, issue)
}
\arguments{
\item{data}{Dataframe containing records which will be affected.}

\item{variables}{Vector of strings (or a single string) containing the names of the variables.}

\item{issue}{String with explanation used for the cleaning log entry.}
}
\value{
Dataframe containing cleaning log entries constructed from `data`.
}
\description{
Set the given variables for the given entries to NA.
}
\examples{
\dontrun{
 recode.set.NA(data = dplyr::filter(raw.main, condition),
 variables = c("question1", "question2"), issue = "explanation")
 }
}
