% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.set.NA.regex}
\alias{recode.set.NA.regex}
\title{Recode a question by setting variables to NA if they are matching a regex pattern.}
\usage{
\method{recode}{set.NA.regex}(
  data,
  variables,
  pattern,
  issue,
  CL_COLS = c("uuid", "uniqui", "loop_index", "variable", "old.value", "new.value",
    "issue")
)
}
\arguments{
\item{data}{Dataframe containing records which will be affected.}

\item{variables}{Vector of strings (or a single string) containing the names of the variables.}

\item{pattern}{Regex pattern which will be used to find entries that will be turned to NA.}

\item{issue}{String with explanation used for the cleaning log entry.}

\item{CL_COLS}{string of columns of the cleaning log defauls to c("uuid","uniqui", "loop_index", "variable", "old.value", "new.value", "issue")}
}
\value{
Dataframe containing cleaning log entries constructed from `data`.
}
\description{
This function is also useful also if you need to simply set some variables to NA - you can put ".*" as the `pattern`.
}
\note{
DO NOT use this function for select_multiple questions. Instead use `recode.multiple.set.NA`
}
\examples{
\dontrun{
 recode.set.NA.regex(data = dplyr::filter(raw.main, condition),
 variables = c("question1", "question2"),
 pattern = "birth_certificates?", issue = "explanation")
 }
}
