% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.set.value.regex}
\alias{recode.set.value.regex}
\title{Recode a question by setting variables to some new value if they are matching a regex pattern.}
\usage{
\method{recode}{set.value.regex}(
  data,
  variables,
  pattern,
  new.value,
  issue,
  affect_na = FALSE,
  CL_COLS = c("uuid", "uniqui", "loop_index", "variable", "old.value", "new.value",
    "issue")
)
}
\arguments{
\item{data}{Dataframe containing records which will be affected.}

\item{variables}{Vector of strings (or a single string) containing the names of the variables.}

\item{pattern}{Regex pattern which will be used to find entries that will be turned to NA.}

\item{new.value}{New value for the recoded rows}

\item{issue}{String with explanation used for the cleaning log entry.}

\item{affect_na}{Whether NA values in the column should be recoded as well.}

\item{CL_COLS}{string of columns of the cleaning log defauls to c("uuid","uniqui", "loop_index", "variable", "old.value", "new.value", "issue")}
}
\value{
Dataframe containing cleaning log entries constructed from `data`.
}
\description{
Recode a question by setting variables to some new value if they are matching a regex pattern.
}
\examples{
\dontrun{
recode.set.value.regex(data = filter(raw.main, condition),
 variables = c("question1", "question2"),
  pattern = "birth_certificates?", issue = "explanation")
  }
}
