% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.sm.elsewhere}
\alias{recode.sm.elsewhere}
\title{Recode elsewhere others}
\usage{
recode.sm.elsewhere(edited.sm, data, tool.survey)
}
\arguments{
\item{edited.sm}{The file with the translated requests loaded with the `load.requests` function}

\item{data}{Dataframe containing Kobo data}

\item{tool.survey}{The survey sheet of your kobo tool}
}
\value{
Returns a cleaning log file
}
\description{
Recode elsewhere others for select multiple questions
}
\examples{
\dontrun{
recode.sm.elsewhere(edited.sm = edited.sm, data = raw.main, tool.survey = tool.survey)
}
}
