% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.so.elsewhere}
\alias{recode.so.elsewhere}
\title{Recode elsewhere others}
\usage{
recode.so.elsewhere(edited.so, data, tool.survey)
}
\arguments{
\item{edited.so}{The file with the translated requests loaded with the `load.requests` function}

\item{data}{Dataframe containing Kobo data}

\item{tool.survey}{The survey sheet of your kobo tool}
}
\value{
Returns a cleaning log file
}
\description{
Recode elsewhere others for select one questions
}
\examples{
\dontrun{
recode.so.elsewhere(edited.so = edited.so, data = raw.main, tool.survey = tool.survey)
}
}
