% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.trans.requests}
\alias{recode.trans.requests}
\title{Create a cleaning log for recoding translated columns}
\usage{
recode.trans.requests(requests, response_col)
}
\arguments{
\item{requests}{The file with the translated requests loaded with the `load.requests` function}

\item{response_col}{The original untranslated response column. "response.uk" by default}
}
\value{
Returns a cleaning log file
}
\description{
Recode translate requests into a cleaning log file
}
\examples{
\dontrun{
recode.trans.requests(requests = requests_file, response_col = "response.uk")
}
}
