% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_extension_cords.R
\name{rm_ex}
\alias{rm_ex}
\title{remove objects from environment}
\usage{
rm_ex(..., envir = globalenv())
}
\arguments{
\item{...}{what objects to retain in environment}

\item{envir}{Which environment to remove objects from, default is the global environment}
}
\description{
Remove all objects from environment EXCEPT
those that are specified.
}
\examples{
\dontrun {
a <- "TEST"
b <- "TEST2"

rm_ex(a)
}
}
