% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save.other.requests}
\alias{save.other.requests}
\title{Save other requests file}
\usage{
save.other.requests(
  df,
  wb_name,
  use_template = F,
  directory,
  template_dir = "resources/other_requests_template.xlsx"
)
}
\arguments{
\item{df}{your other requests dataframe}

\item{wb_name}{The name you'd like your excel file to have}

\item{use_template}{Whether the function should use pre-established template. `False` by default}

\item{directory}{The directory that holds your other requests}

\item{template_dir}{The directory where you've saved your pre-established template if using.
"resources/other_requests_template.xlsx" by default}
}
\value{
Exports a dataframe of other requests into your requests folder
}
\description{
Save other requests file
}
\examples{
\dontrun{
save.other.requests(df = data, wb_name = 'other_requests', directory = 'output/checking/requests/')
}
}
