% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_stats.R
\name{se}
\alias{se}
\title{Standard Error}
\usage{
se(.x)
}
\arguments{
\item{.x}{Variable to calculate the standard error for}
}
\description{
Calculate the standard error of a variable.
Calculated as the standard error over the square root of the
sample size.
}
