% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{select.multiple.check}
\alias{select.multiple.check}
\title{Checks mismatching between cumulative and binary columns for select multiple answers}
\usage{
\method{select}{multiple.check}(data, survey, id_col = "uuid")
}
\arguments{
\item{data}{data frame}

\item{survey}{tool.survey data frame}

\item{id_col}{The name of your uuid column}
}
\value{
Data frame with mismatching entities
}
\description{
Checks mismatching between cumulative and binary columns for select multiple answers
}
\examples{
\dontrun{
select.multiple.check(data=raw.main, survey=tool.survey, id_col="uuid")
}
}
