% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_extension_cords.R
\name{split2}
\alias{split2}
\title{split}
\usage{
split2(x, f, drop = FALSE, nm = NULL, ...)
}
\arguments{
\item{nm}{names to assign the output of \code{split}}
}
\description{
split
}
\examples{
\dontrun{
split2(mtcars, 1:nrow(mtcars), nm = rownames(mtcars))
}
}
