% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{transpose_df}
\alias{transpose_df}
\title{Transpose dataframes, keep colnames and rownames}
\usage{
transpose_df(df)
}
\arguments{
\item{df}{- your dataframe}
}
\value{
- your transposed dataframe
}
\description{
Transpose dataframes, keep colnames and rownames
}
\examples{
library(dplyr)
df <-  data.frame(col1 = 1:2, col2 = 3:4, row.names = letters[1:2])
transpose_df(df)
}
