% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_log_related.R
\name{undo.changes}
\alias{undo.changes}
\title{Undo the things that you've changed in your data}
\usage{
undo.changes(data, clog, is.loop)
}
\arguments{
\item{data}{your dataframe with applied changes}

\item{clog}{your cleaning log}

\item{is.loop}{whether the changes were applied to the loop}
}
\value{
an old version of your dataframe
}
\description{
Provide the same cleaning log as the one that you used to apply the changes and voila...
}
\details{
this function will undo the changes and rewind your data to the previous state! as long as you didn't modify the cleaning log since you applied the changes!
This function basically flips old.value and new.value around in the clog, and applies changes again.
}
\examples{
\dontrun{
undo.changes(data, clog = c_log)
}
}
