% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{utilityR_raw_data}
\alias{utilityR_raw_data}
\title{Raw data}
\usage{
utilityR_raw_data
}
\description{
Survey data
}
\details{
These data sets include HH data (raw and clean data).
}
\examples{
utilityR_raw_data
utilityR_clean_data
utilityR_survey
utilityR_choices
}
\keyword{datasets}
