% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\alias{tool.survey}
\alias{tool.choices}
\title{Test data}
\format{
data.frame object
}
\usage{
data

tool.survey

tool.choices
}
\description{
These data sets are test data collected to serve as examples for the vignettes.
They include examples of clean dataframe `data` and `tool.survey` and `tool.choices`
objects loaded through the `load.tool.choices` and `load.tool.survey` functions
}
\examples{
data
}
\keyword{datasets}
