% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{detect.outliers}
\alias{detect.outliers}
\title{Looking for outliers in the provided data}
\usage{
detect.outliers(df, id, n.sd, method = "o1", is.loop, colnames, ignore_0 = T)
}
\arguments{
\item{df}{Dataframe containing columns which will be observed on outliers}

\item{id}{row's unique identifier}

\item{n.sd}{number of standard deviations will be used in algorithm}

\item{method}{method for outliers detecting: o1, o2, o3, o4.\cr
\code{method = 'o1'} Method based on Z score and logarithmization of the values\cr
\code{method = 'o2'} Modified Z score which based on the median absolute deviation, recommended n.sd\cr
\code{method = 'o3'} Method based on the interquartile range\cr
\code{method = 'o4'} Method based on the median absolute deviation\cr}

\item{is.loop}{Is df a loop data or not}

\item{colnames}{columns which will be observed on outliers}

\item{ignore_0}{ignore 0 values or not, should be TRUE in o2 method, and in another too}
}
\value{
DataFrame containing outliers
}
\description{
Looking for outliers in the provided data
}
\examples{
\dontrun{
raw.main.outliers <- detect.outliers(
 df=raw.main,
 id='uuid',
 colnames=cols.integer_raw.main,
 is.loop=F,
 n.sd=n.sd,
 method="o1",
 ignore_0=T)
}

}
