% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.others}
\alias{recode.others}
\title{Create a cleaning log for recoding other responses}
\usage{
recode.others(
  data,
  or.edited,
  orig_response_col = "response.uk",
  is.loop,
  print_debug = T,
  tool.survey,
  tool.choices,
  label_colname = "label::English"
)
}
\arguments{
\item{data}{Dataframe containing Kobo data}

\item{or.edited}{Dataframe containing the filled-out other_requests file. Needs to contain the standard TEI (True, Existing, Invalid) columns
in one of the following formats: (true.v,existing.v, invalid.v) or (true.other, existing.other, invalid.other),
as well as column `check`. Please use the ```load.requests``` function to load these xlsx files.}

\item{orig_response_col}{The name of the column which stores the original (untranslated!) response for each question.}

\item{is.loop}{Set this to True if the provided dataframe is a loop (it needs to contain loop_index and uuid columns}

\item{print_debug}{Whether debugging information will be printed to screen (about how many responses will be recoded/translated etc.).}

\item{tool.survey}{The survey sheet of your kobo tool}

\item{tool.choices}{The choices sheet of your kobo tool}

\item{label_colname}{The name of your english label column. "label::English" by default}
}
\value{
Dataframe containing cleaning log entries covering recoding others, constructed from `data` and `or.edited`
}
\description{
Use the filled out other_requests file to create cleaning.log.other.
Run this function on every datasheet your data has (main, loop1, loop2, etc...)
Unexpected behavior may occur if some variables from `or.edited` are not present in the data.
}
\examples{
\dontrun{
recode.others(data, or.edited = other_requests_file, orig_response_col = "response.uk",tool.survey=tool.survey,
tool.choices = tool.choices)
}
}
