% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode.others_select_multiple}
\alias{recode.others_select_multiple}
\title{Recode others: function for recoding select_multiple questions}
\usage{
\method{recode}{others_select_multiple}(
  data,
  or.select_multiple,
  orig_response_col = "response.uk",
  print_debug = T,
  is.loop,
  tool.survey_others = tool.survey,
  tool.choices_others = tool.choices,
  label_colname = "label::English",
  CL_COLS = c("uuid", "uniqui", "loop_index", "variable", "old.value", "new.value",
    "issue")
)
}
\arguments{
\item{data}{your dataframe}

\item{or.select_multiple}{Your filled out recoding form after translation and verification (with only select multiple)}

\item{orig_response_col}{The name of your original column with the untranslated responses, "response.uk" by default}

\item{print_debug}{Whether you want the function to print out what it's doing}

\item{is.loop}{Whether you're working with loop data}

\item{tool.survey_others}{The survey sheet of your kobo tool}

\item{tool.choices_others}{The choices sheet of your kobo tool}

\item{label_colname}{The name of your english label column. "label::English" by default}

\item{CL_COLS}{CL_COLS string of columns of the cleaning log. Defauls to c("uuid","uniqui" ,"loop_index", "variable", "old.value", "new.value", "issue")}
}
\value{
A dataframe for the changes that need to be applied to the data
}
\description{
Recode others: function for recoding select_multiple questions
}
\note{
When running loop data, make sure to have a uuid column in your dataframe. This ensures that the resulting tables are uniform
}
\examples{
\dontrun{
recode.others_select_multiple(data, or.select_one = other_requests_file, tool.survey_others=tool.survey,
tool.choices_others = tool.choices)
}
}
